Simplified process kills, now works with images
