+++ src/arcade_station/kill_all_and_reset_pegasus.py
@@ -10,23 +10,10 @@ from arcade_station.core.common.core_functions import *
 from core.common.display_image import *
 from PyQt5.QtWidgets import QApplication
 
-
-PID_FILE = "arcade_station_image.pid"  # Ensure this matches the path in display_image.py
-
-def kill_image_process():
-    if os.path.exists(PID_FILE):
-        with open(PID_FILE, 'r') as f:
-            pid = int(f.read().strip())
-            try:
-                # Use psutil to terminate the process on Windows
-                process = psutil.Process(pid)
-                process.terminate()
-                print(f"Killed process with PID: {pid}")
-            except (psutil.NoSuchProcess, ProcessLookupError):
-                print(f"No process found with PID: {pid}")
-        os.remove(PID_FILE)
-
-# kill_processes_from_toml('processes_to_kill.toml')
-kill_image_process()
-display_image_from_config()
+kill_processes_from_toml('processes_to_kill.toml')
+kill_process_by_identifier("open_image")
+process = launch_script(
+    r'C:/Repositories/arcade_station/src/arcade_station/open_image.py',
+    identifier="open_image"  # This will appear in the process command-line.
+)
 start_pegasus()
